refactor: stores 내부 폴더 분리
